Fix up tests.auth.test_basicauth
Noticed that there were calls to mock.MagicMock.assert_not_called(),
which isn't an actual method and needed to be
self.assertFalse(mock.called). While in there, cleaned up the mocking
out of properties and removed the mock.patch as a context manager,
instead using the decorator on the class itself instead.